Move DSL blocks out of main presenter DSL module for maintainability / legibility.
